home 2024-10-12 08:34

+++ core/src/main/java/com/cac/Temp.java
@@ -12,9 +12,9 @@ public class Temp {
 
     public static class Camera {
         public static class Position {
-            public static long X = 0;
-            public static long Y = 0;
-            public static long Z = 0;
+            public static float X = 0;
+            public static float Y = 0;
+            public static float Z = 0;
         }
     }
 }
